backend/requirements.txt
Django==5.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
virtualenv==20.26.3
requests==2.31.0
